Moved authentication providers from OpenApi package to examples
