--- dotnet/src/Functions/Functions.UnitTests/OpenApi/Authentication/BearerAuthenticationProviderTests.cs
@@ -1,29 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-
-using System;
-using System.Net.Http;
-using System.Threading.Tasks;
-using Microsoft.SemanticKernel.Plugins.OpenApi.Authentication;
-using Xunit;
-
-namespace SemanticKernel.Functions.UnitTests.OpenApi.Authentication;
-
-public class BearerAuthenticationProviderTests
-{
-    [Fact]
-    public async Task AuthenticateRequestAsyncSucceedsAsync()
-    {
-        // Arrange
-        var token = Guid.NewGuid().ToString();
-        using var request = new HttpRequestMessage();
-
-        var target = new BearerAuthenticationProvider(() => Task.FromResult(token));
-
-        // Act
-        await target.AuthenticateRequestAsync(request);
-
-        // Assert
-        Assert.Equal("Bearer", request.Headers.Authorization?.Scheme);
-        Assert.Equal(token, request.Headers.Authorization?.Parameter);
-    }
-}
